requirements.txt
torch
torchvision
jupyter
scikit-learn
scipy
pandas
matplotlib
pillow
tqdm
requests
h5py
pyyaml
langchain
langchain-community
langchain-openai
langchainhub
langchain_experimental
ipykernel
pypdf
chromadb==0.5.3
datasets
sentence-transformers
pydantic
pdfkit
opencv-python
markdown
weasyprint
duckduckgo-search
streamlit
